feat(chat-panel, launchpad, linear-thread): refactor message handling and component structure
- Updated ChatPanel to include onAddMessage prop for enhanced message handling in embedded mode.
- Refactored LaunchPad to pass onAddMessage to ChatPanel, optimizing message context management.
- Introduced LinearThread component to replace ReflyPilot, improving message display and management.
- Updated canvas store to manage linearThreadMessages, enhancing message tracking and retrieval.
- Ensured compliance with code style rules, including the use of single quotes for string literals and performance optimizations.
