+++ packages/ai-workspace-common/src/components/canvas/node-preview/skill-response/enhanced-skill-response.tsx
@@ -1,11 +1,19 @@
-import { memo } from 'react';
+import { memo, useState } from 'react';
 import {
   CanvasNode,
   ResponseNodeMeta,
 } from '@refly-packages/ai-workspace-common/components/canvas/nodes';
-import { ReflyPilot } from '@refly-packages/ai-workspace-common/components/canvas/refly-pilot';
+import { LinearThread } from '@refly-packages/ai-workspace-common/components/canvas/refly-pilot/linear-thread';
 import { cn } from '@refly-packages/utils/cn';
-import { useCanvasStoreShallow } from '@refly-packages/ai-workspace-common/stores/canvas';
+import { useFindThreadHistory } from '@refly-packages/ai-workspace-common/hooks/canvas/use-find-thread-history';
+import { useEffect } from 'react';
+
+interface LinearThreadMessage {
+  id: string;
+  resultId: string;
+  nodeId: string;
+  timestamp: number;
+}
 
 interface EnhancedSkillResponseProps {
   node: CanvasNode<ResponseNodeMeta>;
@@ -15,24 +23,53 @@ interface EnhancedSkillResponseProps {
 
 export const EnhancedSkillResponse = memo(
   ({ node, resultId, className }: EnhancedSkillResponseProps) => {
-    const { getReflyPilotMessagesByResultId } = useCanvasStoreShallow((state) => ({
-      getReflyPilotMessagesByResultId: state.getReflyPilotMessagesByResultId,
-    }));
+    const [messages, setMessages] = useState<LinearThreadMessage[]>([]);
+    const findThreadHistory = useFindThreadHistory();
 
-    // Initialize messages from resultId
-    const reflyPilotMessages = getReflyPilotMessagesByResultId(resultId);
-    console.log('reflyPilotMessages', reflyPilotMessages);
+    // Initialize messages from resultId and its thread history
+    useEffect(() => {
+      if (resultId && node) {
+        // Find thread history based on resultId
+        const threadHistory = findThreadHistory({ resultId });
+
+        // Initialize with empty messages array
+        const initialMessages: LinearThreadMessage[] = [];
+
+        // Add all history nodes to messages
+        const allNodes = [...threadHistory, node];
+
+        allNodes.forEach((historyNode, index) => {
+          const nodeResultId = historyNode?.data?.entityId;
+          if (nodeResultId) {
+            initialMessages.push({
+              id: `history-${historyNode.id}-${index}`,
+              resultId: nodeResultId,
+              nodeId: historyNode.id,
+              timestamp: Date.now() - (allNodes.length - index) * 1000, // Ensure proper ordering
+            });
+          }
+        });
+
+        setMessages(initialMessages);
+      }
+    }, [resultId, node, findThreadHistory]);
+
+    // Handler for adding new messages
+    const handleAddMessage = (message: Omit<LinearThreadMessage, 'timestamp'>) => {
+      setMessages((prev) => [...prev, { ...message, timestamp: Date.now() }]);
+    };
 
     return (
       <div className={cn('flex flex-col h-full w-full', className)}>
         <div className="flex flex-1 overflow-hidden">
-          <ReflyPilot
+          <LinearThread
             resultId={resultId}
             node={node}
             standalone={false}
             className="h-full w-full"
-            initialMessages={reflyPilotMessages}
+            initialMessages={messages}
             useResultIdMapping={true}
+            onAddMessage={handleAddMessage}
           />
         </div>
       </div>
