+++ packages/ai-workspace-common/src/hooks/canvas/index.ts
@@ -23,6 +23,7 @@ export * from './use-action-polling';
 // Context operations
 export * from './use-add-to-context';
 export * from './use-sync-selected-nodes-to-context';
+export * from './use-debounced-context-update';
 
 // Canvas operations
 export * from './use-canvas-sync';
